+++ Near-iOS-UI/changelog.md
@@ -1,3 +1,88 @@
+## 2.11.7 (2019-05-09)
+
+### API Breaking Changes
+
+* None.
+
+### Enhancements
+
+* New method for opening content dialogs immediately.
+* Pass `trackingInfo` to custom JSON taps delegate method.
+
+### Bugfixes
+
+* None.
+
+## 2.11.5 (2019-04-30)
+
+### API Breaking Changes
+
+* None.
+
+### Enhancements
+
+*  `NITNotificationHistoryViewController` delegate has method for detecting tap on custom JSONs.
+
+### Bugfixes
+
+* None.
+
+## 2.11.5 (2019-03-15)
+
+### API Breaking Changes
+
+* None.
+
+### Enhancements
+
+* `NITCouponViewController` and `NITNotificationHistoryViewController` can be used from Interface Builder.
+
+### Bugfixes
+
+* None.
+
+## 2.11.4 (2019-02-20)
+
+### API Breaking Changes
+
+* None.
+
+### Enhancements
+
+* None.
+
+### Bugfixes
+
+* Avoid sending empty comment for Feedback answer.
+
+## 2.11.3 (2019-02-14)
+
+### API Breaking Changes
+
+* None.
+
+### Enhancements
+
+* Automatically track simple notifications as read when rendered in notification list.
+
+### Bugfixes
+
+* None.
+
+## 2.11.2 (2019-02-13)
+
+### API Breaking Changes
+
+None.
+
+### Enhancements
+
+None.
+
+### Bugfixes
+
+* Fixed glitch on simple notifications in history list.
+
 ## 2.10.5 (2018-11-28)
 
 ### API Breaking Changes